requirements.txt
pandas==2.2.2
openpyxl==3.1.2
qrcode==7.4.2
pillow==10.3.0
tqdm == 4.67.1